move MedusaMetrics.et() to its own module
